+++ lucid.js
@@ -42,11 +42,12 @@ function createApp(properties) {
 
         this.currPage.elements[componentName + componentKey] = {
           state: this.components[componentName].state,
+          methods: this.components[componentName].methods,
           dom: child
         };
 
         child.innerHTML = this.components[componentName].render();
-        registerEvents(child, componentName, componentKey);
+        registerDom(child, componentName, componentKey);
       });
     }
   };
@@ -58,27 +59,73 @@ function createApp(properties) {
  * 
  * @param {HTMLElement} element 
  */
-function registerEvents(element, componentName, componentKey) {
+function registerDom(element, componentName, componentKey) {
   element.childNodes.forEach((child) => {
     for (let i = 0; i < child.attributes.length; ++i) {
       const attrib = child.attributes[i];
-      if (attrib.specified && attrib.name.startsWith("on")) {
-        const methodName = child.attributes[i].value;
+
+      if (!attrib.specified)
+        return;
+
+      const attribValue = attrib.value;
+      const result = convertTextVariables(Lucid.app.currPage.elements[componentName + componentKey], attribValue);
+
+      if (attrib.name.startsWith("on")) {
         child.addEventListener(attrib.name.substr(2),
           () => {
-            Lucid.app.components[componentName].methods[methodName]
+            result
               (
                 Lucid.app.currPage.elements[componentName + componentKey].state,
                 (newState) => {
-                  for (const key in newState) {
-                    Lucid.app.currPage.elements[componentName + componentKey].state[key] = newState[key]
-                  }
+                  Lucid.app.currPage.elements[componentName + componentKey].state = newState
+                  console.log(newState);
+
+                  // Re-render the element
+                  const dom = Lucid.app.currPage.elements[componentName + componentKey].dom;
+                  dom.innerHTML = Lucid.app.components[componentName].render();
+                  registerDom(dom, componentName, componentKey);
                 }
               )
           }
         );
-        child.removeAttribute(child.attributes[i].name);
+        child.removeAttribute(attrib.name);
+        --i;
+      }
+      else {
+        attrib.value = result;
       }
     }
+
+    // Convert textContent variables and re-write to the element
+    const result = convertTextVariables(Lucid.app.currPage.elements[componentName + componentKey], child.textContent);
+    child.textContent = result;
   });
+}
+
+function convertTextVariables(obj, text) {
+  let startIndex = 0;
+  let endIndex = 0;
+
+  while (
+    (startIndex = text.indexOf("{{", startIndex)) > -1 &&
+    (endIndex = text.indexOf("}}", startIndex + 2))
+  ) {
+    let variable = text.substring(startIndex + 2, endIndex);
+
+    let properties = variable.split(".");
+
+    let tempObj = obj;
+    for (let i = 0; i < properties.length; ++i)
+      tempObj = tempObj[properties[i]];
+
+    if (typeof tempObj === "function")
+      return tempObj;
+    else
+      text = text.replace("{{" + variable + "}}", tempObj);
+
+    startIndex = 0;
+    endIndex = 0;
+  }
+
+  return text;
 }