+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-### 0.1.0 (January 20, 2021)
+### 0.2.0 (January 20, 2021) 
+---
+- Add ability to re-render when setState is used to change the state
+- Add string variables which can be used in component templates
+- Add string variable support for other non-event HTMLElement attributes and textContent
+- Change access to methods and state inside components
+
+### 0.1.0 (January 20, 2021) 
+---
 - Initial public release